moved shooting to its own script
moved shooting to its own script made it check if it was already
shooting before starting a coroutine so there arent multiple coroutines
for the same enemy shooting
