+++ Assets/Scripts/EnemyGun.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 6f25b1cd2c5d4824b8e291c9627740df
+timeCreated: 1487142721
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
